requirements.txt
asgiref==3.9.1
boto3==1.42.44
botocore==1.42.44
certifi==2025.1.31
charset-normalizer==3.4.0
diff-match-patch==20241021
Django==5.2.6
django-cors-headers==4.7.0
django-extensions==4.1
django-import-export==4.3.9
django-storages==1.14.6
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
gunicorn==23.0.0
idna==3.10
jmespath==1.1.0
Markdown==3.7
packaging==25.0
pillow==11.3.0
psycopg2-binary==2.9.10
pydot==4.0.1
PyJWT==2.9.0
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
requests==2.32.3
s3transfer==0.16.0
setuptools==80.9.0
six==1.17.0
sqlparse==0.5.1
tablib==3.8.0
tzdata==2025.2
urllib3==2.2.2
wheel==0.45.1
